+++ changelog.md
@@ -10,6 +10,7 @@
       - Add ctrl+click selection system
       - Fix code generation of room sounds and smells
       - Add hover previews to display monster look/item longs for room preview and monster preview panels
+      - Add includes to area properties
     - Virtual area editor: Add ctrl+click selection system
 - **Fixed:**
   - Increase progress dialog height